init 13.4.0

+++ tasmota/tasmota_xdrv_driver/xdrv_23_zigbee_9a_upload.ino
@@ -144,7 +144,7 @@ char XModemWaitACK(void)
 {
   char in_char;
   do {
-    int i = 0;
+    uint8_t i = 0;
     while (!ZigbeeSerial->available()) {
       delayMicroseconds(100);
       i++;
